requirements.txt
markdown
pyserial~=3.5
protobuf>=5.26.0
dotmap~=1.3.30
pexpect
pyqrcode~=1.2.1
tabulate~=0.9.0
timeago~=1.0.16
webencodings
requests~=2.32.2
pyparsing
twine
autopep8
pylint
pytest~=8.2.1
pytest-cov
pyyaml~=6.0.1
pytap2~=2.3.0
pdoc3
pypubsub~=4.0.3
bleak~=0.22.1
packaging~=24.0
mypy
mypy-protobuf
types-protobuf
types-tabulate
types-requests
types-setuptools
types-PyYAML
pydantic
setuptools~=69.2.0